Fix MiniCliffEQ crash on non-thread safe FFTW3 routine calls
This fixes the crash when 2 or more instances of MiniCliffEQ is launched, then host calls `OverlapSaveConvolver::setFir()` by resetting or initialization.

On FFTW 3.3.10 manual section 5.4 Thread safety, it says that "the only thread-safe routine in FFTW is fftw_execute (and the new-array variants thereof)." I have no idea if all the VST 3 hosts are initializing plugins in a single thread.