Split common parts from manuals
This change is a refactoring to depricate `rewrite.py`. `rewrite.py` was written as quick hack to rewrite all manuals, when a change affects all plugins in this repository. It was relying on regex, and it was infested by bugs.

There were several attempts to refactor, and following is the conclusion written in my TODO.md.

---

Split components and use template engine. Don't spend too much time about granularity. Coarse grained is fine.

- Everything between download link and Controls section as a component.
- Each controls could be a component (knob, barbox).

Create `pluginname.json` with following elements:

- Version numbers
- Current release name
- Change log
- Old versions