Change BarBox to only edit relevant parameters
Previously, single parameters change was triggering the editing all the other parameters. This resulted in a issue that the automation of single parameter overwrites the automation of other parameters in the same BarBox.

This change ensure that writing the automations of the part of BarBox through GUI do not affect automations of other part of parameters.