Fix crash on Renoise
When opening plugin editor, `PlugEditor::addToControlMap()` was trying to access already released `CControl` instance in `controlMap`. This resulted in use after free, and plugin was crashing.

This commit also includes a change to `ArrayControl::setValueAt()`. Previously, it was assumed that provided parameter ID were contiguous. This resulted in unexpected behaviour in `MatrixKnob`. Thus, `idMap` was added for a case where non-contiguous ID are provided.