Change build process
VST 3 SDK used to be parent of build process. This means that CMake's `project()` is defined inside of VST 3SDK. It felt weird considering that VST 3 SDK acts like a library rather than an application.

Now that weirdness was fixed, and VST 3 SDK 3.7.8 depricated `SMTG_MYPLUGINS_SRC_PATH` option from their CMake, which was used to specify plugin source directory. This commit is made to follow that change.